MULE-8463: Remove code duplication of classes implementing AnnotatedObject interface
