+++ core/src/main/java/org/mule/component/AbstractComponent.java
@@ -6,11 +6,11 @@
  */
 package org.mule.component;
 
+import org.mule.AbstractAnnotatedObject;
 import org.mule.DefaultMuleEvent;
 import org.mule.DefaultMuleMessage;
 import org.mule.OptimizedRequestContext;
 import org.mule.VoidResult;
-import org.mule.api.AnnotatedObject;
 import org.mule.api.MuleContext;
 import org.mule.api.MuleEvent;
 import org.mule.api.MuleException;
@@ -45,10 +45,6 @@ import org.mule.util.ClassUtils;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
-import javax.xml.namespace.QName;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
@@ -56,7 +52,7 @@ import org.apache.commons.logging.LogFactory;
 /**
  * Abstract {@link Component} to be used by all {@link Component} implementations.
  */
-public abstract class AbstractComponent implements Component, MuleContextAware, Lifecycle, AnnotatedObject, MessagingExceptionHandlerAware
+public abstract class AbstractComponent extends AbstractAnnotatedObject implements Component, MuleContextAware, Lifecycle, MessagingExceptionHandlerAware
 {
 
     /**
@@ -72,7 +68,6 @@ public abstract class AbstractComponent implements Component, MuleContextAware, 
     protected MuleContext muleContext;
     protected ComponentLifecycleManager lifecycleManager;
     private MessagingExceptionHandler messagingExceptionHandler;
-    private final Map<QName, Object> annotations = new ConcurrentHashMap<QName, Object>();
 
     public void setMuleContext(MuleContext context)
     {
@@ -353,22 +348,6 @@ public abstract class AbstractComponent implements Component, MuleContextAware, 
             sb.append(System.identityHashCode(this));
         }
         return sb.toString();
-    }
-
-    public final Object getAnnotation(QName name)
-    {
-        return annotations.get(name);
-    }
-
-    public final Map<QName, Object> getAnnotations()
-    {
-        return Collections.unmodifiableMap(annotations);
-    }
-
-    public synchronized final void setAnnotations(Map<QName, Object> newAnnotations)
-    {
-        annotations.clear();
-        annotations.putAll(newAnnotations);
     }
 
     public void setMessagingExceptionHandler(MessagingExceptionHandler messagingExceptionHandler)
