+++ core/src/main/java/org/mule/processor/InvokerMessageProcessor.java
@@ -6,9 +6,9 @@
  */
 package org.mule.processor;
 
+import org.mule.AbstractAnnotatedObject;
 import org.mule.DefaultMuleEvent;
 import org.mule.DefaultMuleMessage;
-import org.mule.api.AnnotatedObject;
 import org.mule.api.MessagingException;
 import org.mule.api.MuleContext;
 import org.mule.api.MuleEvent;
@@ -40,9 +40,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
-import java.util.concurrent.ConcurrentHashMap;
-
-import javax.xml.namespace.QName;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
@@ -56,7 +53,7 @@ import org.apache.commons.logging.LogFactory;
  * methods with the same name and same number of arguments are not supported
  * currently.
  */
-public class InvokerMessageProcessor implements MessageProcessor, Initialisable, MuleContextAware, AnnotatedObject
+public class InvokerMessageProcessor extends AbstractAnnotatedObject implements MessageProcessor, Initialisable, MuleContextAware
 {
     protected final transient Log logger = LogFactory.getLog(getClass());
 
@@ -71,7 +68,6 @@ public class InvokerMessageProcessor implements MessageProcessor, Initialisable,
     protected Method method;
     protected ExpressionManager expressionManager;
     protected MuleContext muleContext;
-    private final Map<QName, Object> annotations = new ConcurrentHashMap<QName, Object>();
 
     @Override
     public void initialise() throws InitialisationException
@@ -357,21 +353,5 @@ public class InvokerMessageProcessor implements MessageProcessor, Initialisable,
     public void setObjectType(Class<?> objectType)
     {
         this.objectType = objectType;
-    }
-
-    public final Object getAnnotation(QName name)
-    {
-        return annotations.get(name);
-    }
-
-    public final Map<QName, Object> getAnnotations()
-    {
-        return Collections.unmodifiableMap(annotations);
-    }
-
-    public synchronized final void setAnnotations(Map<QName, Object> newAnnotations)
-    {
-        annotations.clear();
-        annotations.putAll(newAnnotations);
     }
 }
