+++ core/src/main/java/org/mule/routing/AbstractMatchingRouter.java
@@ -6,9 +6,9 @@
  */
 package org.mule.routing;
 
+import org.mule.AbstractAnnotatedObject;
 import org.mule.DefaultMuleMessage;
 import org.mule.OptimizedRequestContext;
-import org.mule.api.AnnotatedObject;
 import org.mule.api.MessagingException;
 import org.mule.api.MuleEvent;
 import org.mule.api.MuleException;
@@ -22,14 +22,9 @@ import org.mule.api.routing.MatchingRouter;
 import org.mule.api.routing.TransformingMatchable;
 import org.mule.config.i18n.CoreMessages;
 
-import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.CopyOnWriteArrayList;
-
-import javax.xml.namespace.QName;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
@@ -39,7 +34,7 @@ import org.apache.commons.logging.LogFactory;
  * and outbound routers.
  */
 
-public class AbstractMatchingRouter implements MatchingRouter, AnnotatedObject
+public class AbstractMatchingRouter extends AbstractAnnotatedObject implements MatchingRouter
 {
     /**
      * logger used by this class
@@ -49,7 +44,6 @@ public class AbstractMatchingRouter implements MatchingRouter, AnnotatedObject
     protected List<MatchableMessageProcessor> matchableRoutes = new CopyOnWriteArrayList<MatchableMessageProcessor>();
     protected boolean matchAll = false;
     protected MessageProcessor defaultRoute;
-    private final Map<QName, Object> annotations = new ConcurrentHashMap<QName, Object>();
 
     @Override
     public MuleEvent process(MuleEvent event) throws MuleException
@@ -185,24 +179,5 @@ public class AbstractMatchingRouter implements MatchingRouter, AnnotatedObject
                 ((Disposable) route).dispose();
             }
         }
-    }
-
-    @Override
-    public final Object getAnnotation(QName name)
-    {
-        return annotations.get(name);
-    }
-
-    @Override
-    public final Map<QName, Object> getAnnotations()
-    {
-        return Collections.unmodifiableMap(annotations);
-    }
-
-    @Override
-    public synchronized final void setAnnotations(Map<QName, Object> newAnnotations)
-    {
-        annotations.clear();
-        annotations.putAll(newAnnotations);
     }
 }
