+++ core/src/main/java/org/mule/service/AbstractService.java
@@ -6,10 +6,10 @@
  */
 package org.mule.service;
 
+import org.mule.AbstractAnnotatedObject;
 import org.mule.DefaultMuleEvent;
 import org.mule.RequestContext;
 import org.mule.VoidMuleEvent;
-import org.mule.api.AnnotatedObject;
 import org.mule.api.DefaultMuleException;
 import org.mule.api.MessagingException;
 import org.mule.api.MuleContext;
@@ -57,12 +57,6 @@ import org.mule.service.processor.ServiceAsyncRequestReplyRequestor;
 import org.mule.source.ClusterizableMessageSourceWrapper;
 import org.mule.util.ClassUtils;
 
-import java.util.Collections;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
-import javax.xml.namespace.QName;
-
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 
@@ -70,7 +64,7 @@ import org.apache.commons.logging.LogFactory;
  * A base implementation for all Services in Mule
  */
 @Deprecated
-public abstract class AbstractService implements Service, MessageProcessor, AnnotatedObject
+public abstract class AbstractService extends AbstractAnnotatedObject implements Service, MessageProcessor
 {
 
     /**
@@ -112,7 +106,6 @@ public abstract class AbstractService implements Service, MessageProcessor, Anno
 
     protected MessageProcessorChain messageProcessorChain;
     protected MessageInfoMapping messageInfoMapping = new MuleMessageInfoMapping();
-    private final Map<QName, Object> annotations = new ConcurrentHashMap<QName, Object>();
 
     /**
      * Determines the initial state of this service when the model starts. Can be
@@ -773,21 +766,5 @@ public abstract class AbstractService implements Service, MessageProcessor, Anno
         {
             ((Resumable) candidate).resume();
         }
-    }
-
-    public final Object getAnnotation(QName name)
-    {
-        return annotations.get(name);
-    }
-
-    public final Map<QName, Object> getAnnotations()
-    {
-        return Collections.unmodifiableMap(annotations);
-    }
-
-    public synchronized final void setAnnotations(Map<QName, Object> newAnnotations)
-    {
-        annotations.clear();
-        annotations.putAll(newAnnotations);
     }
 }
